refactor: place 컴포넌트 index에 새로 추가된 컴포넌트 추가
