+++ src/app/place/[id]/page.tsx
@@ -1,8 +1,15 @@
 'use client';
 
 import Button from '@/components/common/Button';
-import { DefaultInfo, Header, KagongBox, ReviewBox, Tag, Tooltip } from '@/components/place';
-import ReviewSheet from '@/components/place/ReviewSheet';
+import {
+  DefaultInfo,
+  Header,
+  KagongBox,
+  ReviewBox,
+  ReviewSheet,
+  Tag,
+  Tooltip,
+} from '@/components/place';
 import { useGetPlace } from '@/hooks/queries/place/useGetPlace';
 import Image from 'next/image';
 import { useState } from 'react';
