+++ src/components/place/index.ts
@@ -6,3 +6,6 @@ export { default as ReviewBox } from './ReviewBox';
 export { default as Tag } from './Tag';
 export { default as TimeInfo } from './TimeInfo';
 export { default as Tooltip } from './Tooltip';
+export { default as ReviewSheet } from './ReviewSheet';
+export { default as Tabs } from './Tabs';
+export { default as Rating } from './Rating';
